requirements.txt
alpaca-trade-api==3.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
tensorflow==2.13.0
torch==2.0.1
openai==0.27.8
yfinance==0.2.28
streamlit==1.25.0
plotly==5.15.0
ta-lib==0.4.27
feedparser==6.0.10
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
scipy==1.11.1